requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openpyxl==3.1.5
python-pptx==1.0.2
python-multipart==0.0.9
httpx==0.27.2
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
